ai_latest_development/pyproject.toml
[project]
name = "ai_latest_development"
version = "0.1.0"
description = "ai-latest-development using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.165.1,<1.0.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.46.0",
    "chromadb>=0.5.5",
    "google-generativeai>=0.7.2",
    "pypdf>=4.2.0",
    "streamlit>=1.36.0",
    "reportlab>=4.0.0"
]

[project.scripts]
ai_latest_development = "ai_latest_development.main:run"
run_crew = "ai_latest_development.main:run"
train = "ai_latest_development.main:train"
replay = "ai_latest_development.main:replay"
test = "ai_latest_development.main:test"
ui = "ai_latest_development.ui_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
